sns-testing: check that NNS canisters exist before deploying
Check that PocketIC network has all required subnets and install
NNS-related canisters (core and frontend) only when they're missing.
At this point, we do not check that canisters were correctly
initialized, so this check might not be sufficient to run the scenario
on arbitrary network.
