+++ rs/nervous_system/agent/src/pocketic_impl.rs
@@ -148,7 +148,11 @@ impl CallCanisters for PocketIcAgent<'_> {
     ) -> Result<CanisterInfo, Self::Error> {
         let canister_id = canister_id.into();
 
-        let controllers = self.pocket_ic.get_controllers(canister_id).await;
+        let controllers = self
+            .pocket_ic
+            .try_get_controllers(canister_id)
+            .await
+            .unwrap_or(vec![]);
 
         let Some(controller) = controllers.into_iter().last() else {
             return Err(Self::Error::BlackHole);
