+++ rs/sns/testing/Cargo.toml
@@ -21,9 +21,10 @@ path = "src/lib.rs"
 candid = { workspace = true }
 canister-test = { path = "../../rust_canisters/canister_test" }
 clap = { workspace = true }
+futures = { workspace = true }
 ic-base-types = { path = "../../types/base_types" }
 ic-cdk = { workspace = true }
-ic-management-canister-types = { path = "../../types/management_canister_types" }
+ic-management-canister-types-private = { path = "../../types/management_canister_types" }
 ic-nervous-system-agent = { path = "../../nervous_system/agent" }
 ic-nervous-system-integration-tests = { path = "../../nervous_system/integration_tests" }
 ic-nns-constants = { path = "../../nns/constants" }
