+++ rs/sns/testing/src/lib.rs
@@ -1,1 +1,2 @@
-pub mod pocket_ic;
+pub mod nns_dapp;
+pub mod sns;
